// src/app/aliens/models/ship.ts
export interface Ship {
  id: number;
  size: string;
  name: string;
  maxSpeed: number;
  weaponPower: number;
  weaponRangeKm: number;
  weaponRangeUnits: number;
  damageCapacity: number;
  downedPoints: number;
  destroyedPoints: number;
  escapeTime: number;
  firingInterval: number;
  jpg0: string;
  jpg1: string;
  jpg2: string;
  jpg3: string;
}